refactor: change modules structure
