+++ src/opts/estimate_fee.rs
@@ -1,4 +1,4 @@
-use std::fs;
+use super::transaction::{DeclareTxArgs, DeployAccountTxArgs, DeployTxArgs, InvokeTxArgs};
 
 use clap::Subcommand;
 use eyre::Result;
@@ -8,7 +8,7 @@ use starknet::providers::jsonrpc::models::{
     BroadcastedTransaction, ContractClass,
 };
 
-use super::transaction::{DeclareTxArgs, DeployAccountTxArgs, DeployTxArgs, InvokeTxArgs};
+use std::fs;
 
 #[derive(Debug, Subcommand)]
 pub enum EstimateFeeCommands {
